+++ src/game/mod.rs
@@ -1,13 +1,15 @@
 extern crate dirs;
 
 use crate::character::Character;
-use crate::item::Item;
+use crate::item::{Item, Potion};
 use crate::location::Location;
 use crate::log;
+use crate::quest;
+use crate::quest::QuestList;
 use crate::randomizer::random;
 use crate::randomizer::Randomizer;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::io;
 use tombstone::Tombstone;
 
@@ -27,18 +29,23 @@ pub struct Game {
     pub player: Character,
     pub location: Location,
     pub gold: i32,
+    pub quests: QuestList,
     inventory: HashMap<String, Vec<Box<dyn Item>>>,
     tombstones: HashMap<Location, Tombstone>,
+    inspected: HashSet<Location>,
 }
 
 impl Game {
     pub fn new() -> Self {
+        let quests = QuestList::new();
         Self {
             location: Location::home(),
             player: Character::player(),
             gold: 0,
             inventory: HashMap::new(),
             tombstones: HashMap::new(),
+            inspected: HashSet::new(),
+            quests,
         }
     }
 
@@ -54,19 +61,21 @@ impl Game {
     }
 
     /// Remove the game data and reset this reference.
-    /// Tombstones are preserved across games.
-    pub fn reset(&mut self, hard: bool) {
+    /// Progress is preserved across games.
+    pub fn reset(&mut self) {
         let mut new_game = Self::new();
-
-        if hard {
-            datafile::remove();
-        } else {
-            // preserve tombstones across hero's lifes
-            new_game.tombstones = self.tombstones.drain().collect();
-        }
+        // preserve tombstones and quests across hero's lifes
+        std::mem::swap(&mut new_game.tombstones, &mut self.tombstones);
+        std::mem::swap(&mut new_game.quests, &mut self.quests);
+        // TBD shouldn't chests be preserved?
 
         // replace the current, finished game with the new one
         *self = new_game;
+    }
+
+    /// Recreate the game data, losing all progress.
+    pub fn restet_hard() {
+        datafile::remove();
     }
 
     /// Move the hero's location towards the given destination, one directory
@@ -84,6 +93,32 @@ impl Game {
         Ok(())
     }
 
+    /// Look for chests and tombstones at the current location.
+    /// Remembers previous checks for consistency.
+    pub fn inspect(&mut self) {
+        self.pick_up_tombstone();
+
+        if !self.inspected.contains(&self.location) {
+            self.inspected.insert(self.location.clone());
+
+            // this could be extended to find better items, with a non uniform
+            // probability, and to change according to the distance from home
+            // it's likely better to extract to an item generator module at that point
+            match random().range(6) {
+                0 => {
+                    let gold = random().gold_gained(self.player.level * 200);
+                    log::chest_gold(gold);
+                }
+                1 => {
+                    let potion = Potion::new(self.player.level);
+                    log::chest_item("potion");
+                    self.add_item("potion", Box::new(potion));
+                }
+                _ => {}
+            }
+        }
+    }
+
     /// Set the hero's location to the one given, and apply related side effects.
     pub fn visit(&mut self, location: Location) {
         self.location = location;
@@ -91,7 +126,6 @@ impl Game {
             let recovered = self.player.heal_full();
             log::heal(&self.player, &self.location, recovered);
         }
-        self.pick_up_tombstone();
     }
 
     /// Set the current location to home, and apply related side-effects
@@ -114,6 +148,7 @@ impl Game {
         if let Some(mut items) = self.inventory.remove(&name) {
             if let Some(item) = items.pop() {
                 item.apply(self);
+                quest::handle_item_used(self, &name);
             }
 
             if !items.is_empty() {
@@ -134,12 +169,11 @@ impl Game {
     }
 
     /// If there's a tombstone laying in the current location, pick up its items
-    fn pick_up_tombstone(&mut self) -> bool {
+    fn pick_up_tombstone(&mut self) {
         if let Some(mut tombstone) = self.tombstones.remove(&self.location) {
-            tombstone.pick_up(self);
-            true
-        } else {
-            false
+            let (items, gold) = tombstone.pick_up(self);
+            log::tombstone(&items, gold);
+            quest::handle_tombstone(self);
         }
     }
 
@@ -202,6 +236,7 @@ impl Game {
             let level_up = self.player.add_experience(xp);
 
             log::battle_won(self, xp, level_up, gold);
+            quest::handle_battle_won(self, &enemy, level_up);
             Ok(())
         } else {
             // leave hero items in the location